Hw3

+++ project/link-parser/src/main/java/answers/StackoverflowLinkParserAnswer.java
@@ -1,4 +1,4 @@
 package answers;
 
-public record StackoverflowLinkParserAnswer(Integer id) implements LinkParserAnswer {
+public record StackoverflowLinkParserAnswer(Long id) implements LinkParserAnswer {
 }
